moving get_text_from_ocr under ocr class, including tests + adding requirements.txt
