+++ presidio-image-anonymizer/README.MD
@@ -10,3 +10,8 @@ entities in images.
 Pre-requisites:
 
 * Install [Tesseract OCR](https://github.com/tesseract-ocr/tesseract) by following the instructions on how to install it for your operating system.
+* Install [Spacy model](https://spacy.io/usage/models) to use with the presidio-analyzer.
+
+```sh
+pip install https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.1.0/en_core_web_lg-2.1.0.tar.gz
+```
