+++ presidio-image-anonymizer/presidio_image_anonymizer/ocr.py
@@ -13,3 +13,17 @@ class OCR:
         """
         output_type = pytesseract.Output.DICT
         return pytesseract.image_to_data(image, output_type=output_type)
+
+    @staticmethod
+    def get_text_from_ocr_dict(ocr_result: dict, separator: str = " ") -> str:
+        """Combine the text from the OCR dict to full text.
+
+        :param ocr_result: dictionary containing the ocr results per word
+        :param separator: separator to use when joining the words
+
+        return: str containing the full extracted text as string
+        """
+        if not ocr_result:
+            return ""
+        else:
+            return separator.join(ocr_result["text"])
